requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
pymysql==1.1.0
twilio==8.12.0
python-dotenv==1.0.1
bcrypt==4.1.3
python-jose[cryptography]==3.3.0